requirements.txt
flask
pandas
numpy
scikit-learn
nltk
joblib
newsapi-python
wikipedia
requests
python-dotenv
